Refactoring HostPeerManager
