+++ src/render/peer/AbstractPeerManager.ts
@@ -1,0 +1,23 @@
+import { AbstractMessenger } from "../communications/AbstractMessenger";
+import { AbstractReceiver } from "../communications/AbstractReceiver";
+import { AbstractSignaler } from "./AbstractSignaler";
+
+export abstract class AbstractPeerManager<R extends AbstractReceiver, M extends AbstractMessenger, S extends AbstractSignaler> {
+    private receiver: R;
+    private messenger: M;
+    protected signaler: S;
+
+    constructor(receiver: R, messenger: M, signaler: S) {
+        this.receiver = receiver;
+        this.messenger = messenger;
+        this.signaler = signaler;
+    }
+
+    public getReceiver() {
+        return this.receiver;
+    }
+
+    public getMessenger() {
+        return this.messenger;
+    }
+}
